added $enforced logic to classes

+++ manifests/stig/v79131.pp
@@ -15,15 +15,17 @@ class secure_sqlserver::stig::v79131 (
   Boolean $enforced = false,
   String  $instance = 'MSSQLSERVER',
 ) {
-  $shared_accounts = $facts['sqlserver_shared_accounts']
-  unless $shared_accounts == undef or $shared_accounts == '' {
-    $shared_accounts.each |$drop_user| {
-      $sql_dcl = "DROP USER '${drop_user}';"
-      ::secure_sqlserver::log { "v79131_sql_dcl = \n${sql_dcl}": }
-      sqlserver_tsql{ "remove_shared_account_${drop_user}":
-        instance => $instance,
-        command  => $sql_dcl,
-        require  => Sqlserver::Config[$instance],
+  if $enforced {
+    $shared_accounts = $facts['sqlserver_shared_accounts']
+    unless $shared_accounts == undef or $shared_accounts == '' {
+      $shared_accounts.each |$drop_user| {
+        $sql_dcl = "DROP USER '${drop_user}';"
+        ::secure_sqlserver::log { "v79131_sql_dcl = \n${sql_dcl}": }
+        sqlserver_tsql{ "remove_shared_account_${drop_user}":
+          instance => $instance,
+          command  => $sql_dcl,
+          require  => Sqlserver::Config[$instance],
+        }
       }
     }
   }
